refactor(pagination): renamed some files and functions; Removed unused files
