--- packages/pagination/src/utils/generateUrlForPagination.ts
@@ -1,7 +1,0 @@
-export const generateUrlForPagination =
-  (pathname: string, elementId: string) => (pageNumber: number) => {
-    // const url = new URL(pathname);
-    // url.searchParams.set('p', pageNumber.toString());
-    // return `${url.toString()}#${elementId}`;
-    return 'test';
-  };
